requirements.txt
anthropic>=0.40.0
asyncpg>=0.29.0
pdfplumber>=0.11.0
python-dotenv>=1.0.0
sentence-transformers>=3.0.0
numpy>=1.26.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
